refactor(src): move private functions to dedicated file
